Fix atomic accesses on 32-bit platforms
When atomically accessing 64-bit values on 32-bit platforms, they must
be aligned on 64-bit boundaries, and we're responsible for ensuring
this alignment.

For lint.Stats, we switch to 32-bit fields, so that we don't have to
be careful about how we allocate the struct. 32-bit should be plenty
to keep track of packages and problems.

In the trivial cases, we've moved the 64-bit fields to the top of the
struct.

+++ lint/stats.go
@@ -8,13 +8,13 @@ const (
 )
 
 type Stats struct {
-	State uint64
+	State uint32
 
-	InitialPackages          uint64
-	TotalPackages            uint64
-	ProcessedPackages        uint64
-	ProcessedInitialPackages uint64
-	Problems                 uint64
-	ActiveWorkers            uint64
-	TotalWorkers             uint64
+	InitialPackages          uint32
+	TotalPackages            uint32
+	ProcessedPackages        uint32
+	ProcessedInitialPackages uint32
+	Problems                 uint32
+	ActiveWorkers            uint32
+	TotalWorkers             uint32
 }
